allow use struct field tag in `Exec()` and `Raw()`
