+++ schema/relationship.go
@@ -3,6 +3,7 @@ package schema
 import (
 	"context"
 	"fmt"
+	"gorm.io/gorm/utils"
 	"reflect"
 	"strings"
 
@@ -123,16 +124,17 @@ func (schema *Schema) parseRelation(field *Field) *Relationship {
 }
 
 // User has many Toys, its `Polymorphic` is `Owner`, Pet has one Toy, its `Polymorphic` is `Owner`
-//     type User struct {
-//       Toys []Toy `gorm:"polymorphic:Owner;"`
-//     }
-//     type Pet struct {
-//       Toy Toy `gorm:"polymorphic:Owner;"`
-//     }
-//     type Toy struct {
-//       OwnerID   int
-//       OwnerType string
-//     }
+//
+//	type User struct {
+//	  Toys []Toy `gorm:"polymorphic:Owner;"`
+//	}
+//	type Pet struct {
+//	  Toy Toy `gorm:"polymorphic:Owner;"`
+//	}
+//	type Toy struct {
+//	  OwnerID   int
+//	  OwnerType string
+//	}
 func (schema *Schema) buildPolymorphicRelation(relation *Relationship, field *Field, polymorphic string) {
 	relation.Polymorphic = &Polymorphic{
 		Value:           schema.Table,
@@ -555,7 +557,7 @@ func (rel *Relationship) ParseConstraint() *Constraint {
 	var (
 		name     string
 		idx      = strings.Index(str, ",")
-		settings = ParseTagSetting(str, ",")
+		settings = utils.ParseTagSetting(str, ",")
 	)
 
 	// optimize match english letters and midline
