+++ schema/utils.go
@@ -13,37 +13,6 @@ import (
 
 var embeddedCacheKey = "embedded_cache_store"
 
-func ParseTagSetting(str string, sep string) map[string]string {
-	settings := map[string]string{}
-	names := strings.Split(str, sep)
-
-	for i := 0; i < len(names); i++ {
-		j := i
-		if len(names[j]) > 0 {
-			for {
-				if names[j][len(names[j])-1] == '\\' {
-					i++
-					names[j] = names[j][0:len(names[j])-1] + sep + names[i]
-					names[i] = ""
-				} else {
-					break
-				}
-			}
-		}
-
-		values := strings.Split(names[j], ":")
-		k := strings.TrimSpace(strings.ToUpper(values[0]))
-
-		if len(values) >= 2 {
-			settings[k] = strings.Join(values[1:], ":")
-		} else if k != "" {
-			settings[k] = k
-		}
-	}
-
-	return settings
-}
-
 func toColumns(val string) (results []string) {
 	if val != "" {
 		for _, v := range strings.Split(val, ",") {
